同步主线修复的一些小bug     鸿蒙研究站 | http://weharmonyos.com (国内)               | https://weharmony.github.io (国外)     论坛 | http://bbs.weharmonyos.com     文档中心 | http://open.weharmonyos.com     参考手册 | http://doxygen.weharmonyos.com

+++ apps/lms/src/sample_usr_lms.c
@@ -180,7 +180,8 @@ static void LmsStrcpyTest(void)
         return;
     }
     char *testStr = "bbbbbbbbbbbbbbbbb";
-    printf("[LmsStrcpyTest] strcpy overflow error should be triggered, src string buf size:%d\n", strlen(testStr) + 1);
+    printf("[LmsStrcpyTest] strcpy overflow error should be triggered, src string buf size:%d\n",
+           (int)strlen(testStr) + 1);
     strcpy(buf, testStr);
     free(buf);
     printf("\n-------- LmsStrcpyTest End --------\n");
